mcp-gemini-service/requirements.txt
fastapi
uvicorn[standard]
httpx
google-generativeai
pydantic
pydantic-settings
python-dotenv

